requirements.txt
asgiref==3.2.3
astroid==2.3.3
awsebcli==3.17.1
bcrypt==3.1.7
boto==2.49.0
botocore==1.14.17
cement==2.8.2
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
colorama==0.3.9
configparser==3.5.0
cryptography==2.8
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.39.1
django-appconf==1.0.3
django-cors-headers==3.2.0
django-filebrowser==3.11.1
django-grappelli==2.12.4
django-hitcount==1.3.1
django-imagekit==4.0.2
django-six==1.0.4
django-sslserver==0.20
djangorestframework==3.11.0
docutils==0.14
fabric==2.5.0
future==0.16.0
httpie==1.0.3
idna==2.7
invoke==1.4.1
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
paramiko==2.7.1
pathspec==0.5.9
pilkit==2.0
Pillow==6.1.0
prettytable==0.7.2
pycparser==2.19
pycryptodome==3.6.1
Pygments==2.4.2
pylint==2.4.4
PyNaCl==1.3.0
pypiwin32==223
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
pywin32==224
PyYAML==5.2
requests==2.20.1
requests-oauthlib==1.2.0
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.0
termcolor==1.1.0
typed-ast==1.4.0
urllib3==1.24.3
wcwidth==0.1.7
whitenoise==4.1.3
wrapt==1.11.2
